Backend/api-services/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
cassandra-driver==3.29.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.11
geomet==0.2.1.post1
h11==0.14.0
h3==4.2.2
idna==3.10
kafka-python-ng==2.2.3
minio==7.2.0
passlib==1.7.4
opencv-python
pillow==11.1.0
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.4.0
python-json-logger==2.0.7
python-multipart==0.0.20
redis==5.2.1
rsa==4.9
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.34.0
geographiclib
websockets==15.0.1
redis==5.2.1
moviepy
ffmpeg-python